Increase rate decimal places
This seems the max number of decimal places that openexchangerates returns. Maybe South migrations or similar would be useful to manage schema changes?

Related: the openexchangerates Python library seems to convert the JSON value to a float, so accuracy may be lost in its conversion back to a decimal here.